// src/app/api-v4/model/v4ListItemInput.ts
/**
 * TMDB API v4 Lists
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { V4ListMediaType } from './v4ListMediaType';


export interface V4ListItemInput { 
    media_type: V4ListMediaType;
    media_id: number;
}
export namespace V4ListItemInput {
}