// src/app/api-v4/model/v4ListItemStatus.ts
/**
 * TMDB API v4 Lists
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface V4ListItemStatus { 
    success?: boolean;
    status_code?: number;
    status_message?: string;
    id?: string;
    item_present?: boolean;
}