// src/app/api-v4/model/v4ListSortBy.ts
/**
 * TMDB API v4 Lists
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const V4ListSortBy = {
    OriginalOrderAsc: 'original_order.asc',
    OriginalOrderDesc: 'original_order.desc',
    VoteAverageAsc: 'vote_average.asc',
    VoteAverageDesc: 'vote_average.desc',
    PrimaryReleaseDateAsc: 'primary_release_date.asc',
    PrimaryReleaseDateDesc: 'primary_release_date.desc',
    TitleAsc: 'title.asc',
    TitleDesc: 'title.desc'
} as const;
export type V4ListSortBy = typeof V4ListSortBy[keyof typeof V4ListSortBy];