// src/app/api-v4/model/v4StatusResponse.ts
/**
 * TMDB API v4 Lists
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface V4StatusResponse { 
    success?: boolean;
    status_code?: number;
    status_message?: string;
}